moving getNormalizedRelativePath to existing util class
