--- plugins/amazonq/codewhisperer/jetbrains-community/src/software/aws/toolkits/jetbrains/services/codewhisperer/codescan/utils/PathUtils.kt
@@ -1,12 +1,0 @@
-// Copyright 2025 Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-package software.aws.toolkits.jetbrains.services.codewhisperer.codescan.utils
-
-import java.nio.file.Path
-import java.nio.file.Paths
-
-object PathUtils {
-    fun getNormalizedRelativePath(projectName: String, relativePath: Path): String =
-        Paths.get(projectName).resolve(relativePath).normalize().toString()
-}
