+++ plugins/amazonq/codewhisperer/jetbrains-community/src/software/aws/toolkits/jetbrains/services/codewhisperer/util/CodeWhispererUtil.kt
@@ -56,6 +56,8 @@ import software.aws.toolkits.resources.message
 import software.aws.toolkits.telemetry.CodewhispererCompletionType
 import software.aws.toolkits.telemetry.CodewhispererGettingStartedTask
 import software.aws.toolkits.telemetry.CredentialSourceId
+import java.nio.file.Path
+import java.nio.file.Paths
 
 // Controls the condition to send telemetry event to CodeWhisperer service, currently:
 // 1. It will be sent for Builder ID users, only if they have optin telemetry sharing.
@@ -329,6 +331,9 @@ object CodeWhispererUtil {
             WindowManager.getInstance().setAlphaModeRatio(it, alpha)
         }
     }
+
+    fun getNormalizedRelativePath(projectName: String, relativePath: Path): String =
+        Paths.get(projectName).resolve(relativePath).normalize().toString()
 }
 
 enum class CaretMovement {
